반복적 메서드 함수파라미터화 리팩토링
