+++ src/main/java/com/example/movies/service/makeuri/MakeUri.java
@@ -1,0 +1,11 @@
+package com.example.movies.service.makeuri;
+
+import com.example.movies.dto.SearchReq;
+
+import java.net.URI;
+
+public interface MakeUri {
+
+    public URI makeURI(SearchReq searchReq);
+
+}
