+++ src/main/java/com/example/movies/service/makeuri/UriById.java
@@ -1,0 +1,12 @@
+package com.example.movies.service.makeuri;
+
+import com.example.movies.dto.SearchReq;
+
+import java.net.URI;
+
+public class UriById implements MakeUri{
+    @Override
+    public URI makeURI(SearchReq searchReq) {
+        return null;
+    }
+}
